Solve flaky test on PartitionedJournalSpecTestCases.
With the addition of triggers that can refuse an insert on the journal, concurrent inserts on the journal for the same persistence_id can fail due to sequence number being smaller then what is already recorded. Therefore, not having gaps among ordering can only be verified across all persistence_ids and not a single one.